requirements.txt
pyautogen[gemini]==0.2.35
python-dotenv
psycopg2-binary
streamlit
pandas
tabulate